refactor(process-table): split the snapshot module along the capture/index seam
`src/shared/process-table-snapshot.ts` counted 308 lines against the 300-line
`max-lines` cap, so `main` itself fails the static-analysis lane for every PR.

The file already held two clusters with different reasons to change: the POSIX
`ps` capture layer (argv, buffer ceiling, both parsers, the truncation guard,
and the in-flight/TTL snapshot reader) and the derived correlation indexes,
which are generic over `ProcessIdentityRow` and shared with the Windows
process table that never forks `ps`. The index cluster moves to
`src/shared/process-table-index.ts` verbatim, together with the types it
exclusively owns (`ProcessTableIndexStats`, `ProcessIdentityRow`,
`ProcessTableIndexOf`, `ProcessTableIndex`).

Import direction is one-way: the index module takes a type-only
`ProcessTableRow`, and the capture module imports nothing back.

Pure refactor — same exports, same semantics, no suppression added.
