+++ src/main/providers/agent-foreground-process-batch.test.ts
@@ -1,9 +1,9 @@
 import { describe, expect, it } from 'vitest'
 import {
   buildProcessTableIndex,
-  parseStrictProcessTableRows,
   type ProcessTableIndexStats
-} from '../../shared/process-table-snapshot'
+} from '../../shared/process-table-index'
+import { parseStrictProcessTableRows } from '../../shared/process-table-snapshot'
 import {
   resolveAgentForegroundProcessesBatch,
   resolveAgentForegroundProcessesFromIndex
